Move platform-specific functions to os*.cpp
