+++ src/flightRecorder.cpp
@@ -22,50 +22,15 @@
 #include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
-#include <sys/time.h>
 #include <sys/types.h>
 #include <unistd.h>
 #include "flightRecorder.h"
+#include "os.h"
 #include "profiler.h"
 #include "vmStructs.h"
 
 
 #define ARRAY_SIZE(arr)  (sizeof(arr) / sizeof(arr[0]))
-
-#ifdef __APPLE__
-
-#include <mach/mach_time.h>
-#include <libkern/OSByteOrder.h>
-
-static mach_timebase_info_data_t timebase = {0, 0};
-
-static inline u64 pd_nanotime() {
-    if (timebase.denom == 0) {
-        mach_timebase_info(&timebase);
-    }
-    return (u64)mach_absolute_time() * timebase.numer / timebase.denom;
-}
-
-static inline u64 pd_hton64(u64 x) {
-    return OSSwapHostToBigInt64(x);
-}
-
-#else // !__APPLE__
-
-#include <byteswap.h>
-
-static inline u64 pd_nanotime() {
-    struct timespec tp;
-    clock_gettime(CLOCK_MONOTONIC, &tp);
-    return (u64)tp.tv_sec * 1000000000 + tp.tv_nsec;
-}
-
-static inline u64 pd_hton64(u64 x) {
-    return htonl(1) == 1 ? x : bswap_64(x);
-}
-
-#endif // __APPLE__
-
 
 const int RECORDING_BUFFER_SIZE = 65536;
 const int RECORDING_LIMIT = RECORDING_BUFFER_SIZE - 4096;
@@ -293,7 +258,7 @@ class Buffer {
     }
 
     void put64(u64 v) {
-        *(u64*)(_data + _offset) = pd_hton64(v);
+        *(u64*)(_data + _offset) = OS::hton64(v);
         _offset += 8;
     }
 
@@ -338,16 +303,16 @@ class Recording {
 
   public:
     Recording(int fd) : _fd(fd), _bytes_written(0), _symbol_map(), _class_map(), _method_map() {
-        _start_time = millis();
-        _start_nanos = nanotime();
+        _start_time = OS::millis();
+        _start_nanos = OS::nanotime();
 
         writeHeader(_buf);
         flush(_buf);
     }
 
     ~Recording() {
-        _stop_nanos = nanotime();
-        _stop_time = millis();
+        _stop_nanos = OS::nanotime();
+        _stop_time = OS::millis();
 
         for (int i = 0; i < CONCURRENCY_LEVEL; i++) {
             flush(&_buf[i]);
@@ -370,21 +335,11 @@ class Recording {
         (void)result;
 
         // Patch metadata offset
-        metadata_offset = pd_hton64(metadata_offset);
+        metadata_offset = OS::hton64(metadata_offset);
         result = pwrite(_fd, &metadata_offset, sizeof(metadata_offset), 8);
         (void)result;
 
         close(_fd);
-    }
-
-    u64 millis() {
-        struct timeval tv;
-        gettimeofday(&tv, NULL);
-        return (u64)tv.tv_sec * 1000 + tv.tv_usec / 1000;
-    }
-
-    u64 nanotime() {
-        return pd_nanotime();
     }
 
     int lookup(std::map<std::string, int>& map, std::string key) {
@@ -736,7 +691,7 @@ class Recording {
         Buffer* buf = &_buf[lock_index];
         buf->put32(30);
         buf->put32(EVENT_EXECUTION_SAMPLE);
-        buf->put64(nanotime());
+        buf->put64(OS::nanotime());
         buf->put32(tid);
         buf->put64(call_trace_id);
         buf->put16(STATE_RUNNABLE);
